requirements.txt
PyQt6==6.7.1
openai==1.43.0
anthropic==0.49.0
torch==2.6.0
transformers==4.51.3
accelerate==1.6.0
requests==2.32.3
python-dotenv==1.0.1
cryptography==42.0.2
keyring==24.2.0 